Errata
I might be a bit late, but here are a few more corrections. I was confused when I read "Finally, we propose some research challenges: p 1. " on the website, but looking at the source, I suspect that the "p" is an error, and the "1" should be the first list item, so I made that change. 

After having read the entire article, it reminds me a lot of the model-driven development movement that was popular when I was in university. It sounded very promising at the time, but not much has come of it many years later. The only feedback I can offer on the content itself is that while it offers a grand vision, it doesn't address how to achieve that vision in any concrete way, and it leaves me with many questions. I guess I will have to wait for future articles to find out how domain models can be represented in a dynamic, general, and yet simple and useful way!